refactor: extract dateUtils + stringUtils, dedupe renderMarkdown
- dateUtils: normalizeDate moved from FeaturedContent (single caller was fine,
  but now reusable for content sorting if needed later)
- stringUtils: renderMarkdown moved from ContentList + Projects (identical dupes,
  now centralized with getCtaButtonClasses from themeClassHelper)
- FeaturedContent: imports normalizeDate from utils
- ContentList: imports renderMarkdown from utils, removed local duplicate
- Projects: same treatment
