+++ src/components/FeaturedContent.jsx
@@ -5,6 +5,7 @@ import { useSubstack } from '../hooks/useSubstack';
 import { getYouTubeVideo } from '../utils/contentful';
 import { getLatestYouTubeVideo } from '../utils/youtube';
 import { getLatestInstagramPost } from '../utils/instagram';
+import { normalizeDate } from '../utils/dateUtils';
 import Loading from './Loading';
 import YouTubeCard from './YouTubeCard';
 import SubstackCard from './SubstackCard';
@@ -12,35 +13,6 @@ import InstagramCard from './InstagramCard';
 import fallbackVideoData from '../dev-data/youtubeVideo.json';
 import fallbackPostData from '../dev-data/featuredPost.json';
 import fallbackInstagramData from '../dev-data/instagramPost.json';
-
-/**
- * Normalizes various date formats to a comparable Date object
- * @param {string} dateStr - Date string in various formats
- * @returns {Date|null} Parsed Date or null if invalid
- */
-const normalizeDate = dateStr => {
-  if (!dateStr) return null;
-
-  // Try parsing as ISO 8601 first (YouTube format)
-  let date = new Date(dateStr);
-  if (!isNaN(date.getTime())) {
-    return date;
-  }
-
-  // Try parsing RSS format (Substack: "Mon, 06 Jan 2025 11:00:00 GMT")
-  // The Date constructor should handle this, but let's be explicit
-  const rssMatch = dateStr.match(
-    /^(\w{3}),\s+(\d{1,2})\s+(\w{3})\s+(\d{4})\s+(\d{2}:\d{2}:\d{2})\s*(\w+)?$/
-  );
-  if (rssMatch) {
-    date = new Date(dateStr);
-    if (!isNaN(date.getTime())) {
-      return date;
-    }
-  }
-
-  return null;
-};
 
 /**
  * FeaturedContent component
